requirements.txt

pyspark
pandas
faker
sqlalchemy
psycopg2-binary
apache-airflow
plotly
dash
requests
